# Fix performance regression in reading from the request body