### What does this PR do?

Since we execute microtasks at the end of each request instead of all requests now, we also have to drain the microtasks after any of the callbacks from uWS

### How did you verify your code works?

After:
```js
Summary:
  Success rate:	1.0000
  Total:	1.6464 secs
  Slowest:	0.0056 secs
  Fastest:	0.0000 secs
  Average:	0.0004 secs
  Requests/sec:	121475.3116

  Total data:	0 B
  Size/request:	0 B
  Size/sec:	0 B

Response time histogram:
  0.000 [170]    |
  0.000 [7184]   |■■
  0.000 [100164] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.000 [49522]  |■■■■■■■■■■■■■■■
  0.001 [25150]  |■■■■■■■■
  0.001 [7917]   |■■
  0.001 [4490]   |■
  0.001 [2155]   |
  0.001 [1811]   |
  0.001 [330]    |
  0.001 [1107]   |

Latency distribution:
  10% in 0.0003 secs
  25% in 0.0003 secs
  50% in 0.0004 secs
  75% in 0.0005 secs
  90% in 0.0006 secs
  95% in 0.0007 secs
  99% in 0.0010 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0020 secs, 0.0014 secs, 0.0028 secs
  DNS-lookup:	0.0000 secs, 0.0000 secs, 0.0003 secs

Status code distribution:
  [200] 200000 responses
```

Before (Bun v0.8)
```js
Summary:
  Success rate:	1.0000
  Total:	9.0704 secs
  Slowest:	0.0575 secs
  Fastest:	0.0000 secs
  Average:	0.0023 secs
  Requests/sec:	22049.6255

  Total data:	0 B
  Size/request:	0 B
  Size/sec:	0 B

Response time histogram:
  0.001 [4783]   |■
  0.001 [121170] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.002 [16151]  |■■■■
  0.002 [10845]  |■■
  0.003 [16191]  |■■■■
  0.004 [5689]   |■
  0.004 [4450]   |■
  0.005 [2868]   |
  0.006 [2394]   |
  0.006 [1539]   |
  0.007 [13920]  |■■■

Latency distribution:
  10% in 0.0007 secs
  25% in 0.0009 secs
  50% in 0.0010 secs
  75% in 0.0023 secs
  90% in 0.0045 secs
  95% in 0.0083 secs
  99% in 0.0177 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0051 secs, 0.0043 secs, 0.0060 secs
  DNS-lookup:	0.0001 secs, 0.0000 secs, 0.0005 secs

Status code distribution:
  [200] 200000 responses
```

Code:

```js
Bun.serve({
  port: 3000,
  async fetch(req) {
    await req.json();
    return new Response();
  },
});
```